Splitting the two parent-child components in the InternationalizationTab
Therefore, i18n and the TranslationTableRow have been split into two different components and the tests for each have been added
